Data.Array.Accelerate.Numeric.LinearAlgebra.BLAS.Level2

Description

Level 2 (matrix-vector) BLAS operations.

Documentation

class (Elt a, Num a) => Numeric a

Minimal complete definition

numericR

type Vector = Array DIM1

Vectors are one-dimensional arrays

type Matrix = Array DIM2

Matrices are two-dimensional arrays

data Transpose

Many operations allow you to implicitly transpose the arguments. For a given input matrix mat with dimensions Z :. m :. n (that is; m rows and n columns):

Constructors

N
Leave the matrix as is.

T
Treat the matrix as implicitly transposed, with dimensions Z :. n :. m. Entry Z :. j :. i is treated as actually being entry Z :. i :. j.

H
Implicitly transpose and conjugate the input matrix. For complex-valued matrices a given element mat ! Z:.j:.i == x :+ y will be treated as actually being mat ! Z:.i:.j == x :+ (-y).

gemv

Arguments

:: Numeric e

=> Exp e
\( \alpha \)

-> Transpose
Operation to apply to A

-> Acc (Matrix e)
A

-> Acc (Vector e)
x

-> Acc (Vector e)
y

Computes the matrix-vector product of a general matrix.

\[ y = \alpha * \mathrm{op}(A) * x \]

where:

shape \(\mathrm{op}(A)\) = Z :. m :. n

shape \(x\) = Z :. n

shape \(y\) = Z :. m

https://software.intel.com/en-us/mkl-developer-reference-c-cblas-gemv