Data.Array.Accelerate.Numeric.Sum

Description

Functions for summing floating point numbers more accurately than the straightforward sum operation.

In the worst case, the sum function accumulates error at a rate proportional to the number of values being summed. The algorithms in this module implement different methods of compensated summation, which reduce the accumulation of numeric error so that it grows much more slowly than the number of inputs (e.g. logarithmically), or remains constant.

Summation type class

class (Elt a, Elt (s a)) => Summation s a where

A class for the summation of floating-point numbers

Methods

add :: Exp (s a) -> Exp (s a) -> Exp (s a)

Add a value to the sum

zero :: Exp (s a)

The identity of the summation

into :: Proxy s -> Exp a -> Exp (s a)

Insert a value into the summation

from :: Proxy s -> Exp (s a) -> Exp a

Summarise the result of summation

sum :: (Summation s a, Shape sh) => Proxy s -> Acc (Array (sh :. Int) a) -> Acc (Array sh a)

Sum an array using a particular compensation scheme.

>>> let xs = [1.0, 1.0e100, 1.0, -1.0e100] :: [Double]
>>> Prelude.sum xs
0.0

>>> let ys = fromList (Z:.4) [1.0, 1.0e100, 1.0, -1.0e100] :: Vector Double
>>> sum kbn (use ys)
Scalar Z [2.0]

Kahan-Babuška-Neumaier summation

data KBN a

Kahan-Babuška-Neumaier summation. This is a little more computationally costly than plain Kahan summation, but is always at least as accurate.

Constructors

KBN a a

kbn :: Proxy KBN

Return the result of a Kahan-Babuška-Neumaier sum.

Order-2 Kahan-Babuška summation

data KB2 a

Second-order Kahan-Babuška summation. This is more computationally costly than Kahan-Babuška-Neumaier summation. Its advantage is that it can lose less precision (in admittedly obscure cases).

This method compensates for error in both the sum and the first-order compensation term, hence the use of "second order" in the name.

Constructors

KB2 a a a

kb2 :: Proxy KB2

Return the result of a second-order Kahan-Babuška sum.

Kahan summation

data Kahan a

Kahan summation. This is the least accurate of the compensated summation methods. This summation method is included only for completeness.

Constructors

Kahan a a

kahan :: Proxy Kahan

Return the result of a Kahan sum.