Data.Array.Accelerate.Numeric.LinearAlgebra.BLAS.Level1

Description

Level 1 (vector-vector) BLAS operations.

Documentation

class (Elt a, Num a) => Numeric a

Minimal complete definition

numericR

type Vector e = Array DIM1 e

Vectors are one-dimensional arrays

sdot :: forall e. Numeric e => Exp e -> Acc (Vector e) -> Acc (Vector e) -> Acc (Scalar e)

Computes a vector-vector dot product, using double precision accumulation of the intermediate result. Includes a scalar (initial) value to be added to the inner product.

https://software.intel.com/en-us/mkl-developer-reference-c-cblas-sdot

dotu :: Numeric e => Acc (Vector e) -> Acc (Vector e) -> Acc (Scalar e)

Computes a vector-vector dot product

\[ res = \sum_i x_i * y_i \]

https://software.intel.com/en-us/mkl-developer-reference-c-cblas-dotu

dotc :: forall e. Numeric (Complex e) => Acc (Vector (Complex e)) -> Acc (Vector (Complex e)) -> Acc (Scalar (Complex e))

Computes a dot product of a conjugated vector with another vector

\[ res = \sum_i \mathrm{conj}(x_i) * y_i \]

https://software.intel.com/en-us/mkl-developer-reference-c-cblas-dotc

asum :: forall e. Numeric e => Acc (Vector e) -> Acc (Scalar (NumericBaseT e))

Computes the sum of magnitudes of the vector elements. For complex values, this is given by \(\sum_i \|\mathrm{real}(x_i)\| + \|\mathrm{imag}(x_i)\|\).

https://software.intel.com/en-us/mkl-developer-reference-c-cblas-asum

amax :: forall e. Numeric e => Acc (Vector e) -> Acc (Scalar Int)

Return the index of the element with the maximum absolute value.

https://software.intel.com/en-us/mkl-developer-reference-c-cblas-i-amax

amin :: forall e. Numeric e => Acc (Vector e) -> Acc (Scalar Int)

Return the index of the element with the minimum absolute value.

https://software.intel.com/en-us/mkl-developer-reference-c-cblas-i-amin